modules: split utility
